// src/app/model/stageQRQC.ts
export class StageQRQC {
    id: number;
    dateHeure: Date;
    signalePar: string;
    departementLigne: string;
    descriptionProbleme: string;
    gravite: string;
    actionImmediate: string;
    analyseCauseRacine: string;
    actionsCorrectives: string;
    personneResponsable: string;
    dateLimiteActions: Date;
    verificationActions: string;
    actionsSuivi: string;
    statut: string;
    commentairesNotes: string;
  }
  